fix: 크루 리뷰 피드백 반영

+++ api/crimp-core/src/main/java/io/crimp/core/repository/crew/CrewSearchRow.java
@@ -18,7 +18,7 @@ public record CrewSearchRow(
         CrewLevelBand levelBand,
         CrewStyle style,
         CrewJoinPolicy joinPolicy,
-        Integer capacity,
+        Short capacity,
         Integer memberCount,
         Instant createdAt,
         String homeGymExtId,
